backend/requirements.txt
fastapi
uvicorn
google-genai
pydantic
python-dotenv
pydantic[email] 
passlib[bcrypt] 
python-jose
supabase
python-multipart
requests
# transformers
# torch